fix: Match 및 User entity 수정, 싱글플레이 로직 개선
- Match 및 User 엔티티에서 `bigint` → `int4`로 타입 변경
- preparing 단계 진입 시 이전 세션 정보 초기화 로직 추가
- QuestionProvider 제거 및 QuestionCtx 구조 통합으로 코드 단순화

+++ packages/backend/src/match/entity/match.entity.ts
@@ -13,16 +13,16 @@ import { UserProblemBank } from '../../problem-bank/entity';
 
 @Entity('matches')
 export class Match {
-  @PrimaryGeneratedColumn('increment', { type: 'bigint' })
+  @PrimaryGeneratedColumn('increment', { type: 'int4' })
   id: number;
 
-  @Column({ type: 'bigint', nullable: false, name: 'player1_id' })
+  @Column({ type: 'int4', nullable: false, name: 'player1_id' })
   player1Id: number;
 
-  @Column({ type: 'bigint', nullable: true, name: 'player2_id' })
+  @Column({ type: 'int4', nullable: true, name: 'player2_id' })
   player2Id: number | null;
 
-  @Column({ type: 'bigint', nullable: true, name: 'winner_id' })
+  @Column({ type: 'int4', nullable: true, name: 'winner_id' })
   winnerId: number | null;
 
   @Column({
